refactor diagnosis manager
