+++ dlrover/python/diagnosis/common/diagnostician.py
@@ -13,6 +13,13 @@
 
 from abc import ABCMeta
 
+from dlrover.python.common.log import default_logger as logger
+from dlrover.python.diagnosis.common.diagnosis_action import (
+    DiagnosisAction,
+    EventAction,
+    NoAction,
+)
+
 
 class DiagnosisObservation(metaclass=ABCMeta):
     """
@@ -27,3 +34,38 @@ class DiagnosisObservation(metaclass=ABCMeta):
     @property
     def observation(self):
         return self._observation
+
+
+class Diagnostician:
+    """
+    Diagnostician is to observe problems and resolve those problems.
+
+    It includes three APIs here:
+    1. observe: observe if one particular problem happened or not
+    2. resolve: generates the DiagnosisAction to handle the problem
+    observed.
+    3. diagnose: define the procedure of the whole diagnosis for
+    a particular problem.
+    """
+
+    def __init__(self):
+        pass
+
+    def observe(self, **kwargs) -> DiagnosisObservation:
+        # observe if particular problem happened
+        return DiagnosisObservation("unknown")
+
+    def resolve(
+        self, problem: DiagnosisObservation, **kwargs
+    ) -> DiagnosisAction:
+        # explore the solution to resolve the problem
+        return EventAction()
+
+    def diagnose(self, **kwargs) -> DiagnosisAction:
+        # define the diagnosis procedure
+        try:
+            ob = self.observe(**kwargs)
+            return self.resolve(ob, **kwargs)
+        except Exception as e:
+            logger.error(f"Fail to diagnose the problem: {e}")
+            return NoAction()
