+++ dlrover/python/diagnosis/diagnostician/failure_node_diagnostician.py
@@ -13,11 +13,13 @@
 
 from dlrover.python.common.log import default_logger as logger
 from dlrover.python.diagnosis.common.constants import Observation
-from dlrover.python.diagnosis.common.diagnostician import DiagnosisObservation
+from dlrover.python.diagnosis.common.diagnostician import (
+    DiagnosisObservation,
+    Diagnostician,
+)
 from dlrover.python.diagnosis.datacollector.training_log_collector import (
     TrainingLogCollector,
 )
-from dlrover.python.diagnosis.diagnostician.diagnostician import Diagnostician
 
 
 class FailureNodeDiagnostician(Diagnostician):
